+++ src/components/new-chart/header.tsx
@@ -1,6 +1,5 @@
 import {
   Box,
-  Button,
   FormControl,
   InputLabel,
   MenuItem,
@@ -10,46 +9,24 @@ import {
   TextField,
   Typography,
 } from "@mui/material";
-import React, { FormEvent, useState } from "react";
+import React, { useState } from "react";
 
-interface Props {
-  onSubmit: (value: FormDataEntryValue) => void;
-  onCancel: () => void;
-}
 
-type ChartCategory = "Select" | "Topology" | "Span";
+type ChartCategory = "Select" | "Topology" | "Trace";
 
 const isChartCategory = (value: string): value is ChartCategory => {
   switch (value) {
     case "Select":
     case "Topology":
-    case "Span":
+    case "Trace":
       return true;
     default:
       return false;
   }
 };
 
-const Header: React.VFC<Props> = ({ onSubmit, onCancel }) => {
-  const [isValidInputs, setIsValidInputs] = useState<Boolean | null>(null);
+const Header: React.VFC = () => {
   const [chartCategory, setChartCategory] = useState<ChartCategory>("Select");
-
-  const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setIsValidInputs(true);
-    const data = new FormData(event.currentTarget);
-    const chartName = data.get("chartName");
-    if (chartName) {
-      onSubmit(chartName);
-    } else {
-      setIsValidInputs(false);
-    }
-  };
-
-  const handleCancel = (event: React.MouseEvent) => {
-    event.preventDefault();
-    onCancel();
-  };
 
   const handleChartCategoryChange = (event: SelectChangeEvent) => {
     const data = event.target.value;
@@ -68,12 +45,10 @@ const Header: React.VFC<Props> = ({ onSubmit, onCancel }) => {
         flexWrap: "wrap",
         height: "3rem",
         width: "auto",
-        px: 4,
         bgcolor: "#8bc34a",
         justifyContent: "space-between",
       }}
       component="form"
-      onSubmit={handleSubmit}
     >
       <Stack direction="row" spacing={2} alignItems="center" flexWrap="wrap">
         <TextField
@@ -96,22 +71,9 @@ const Header: React.VFC<Props> = ({ onSubmit, onCancel }) => {
               <em>select</em>
             </MenuItem>
             <MenuItem value="Topology">Topology</MenuItem>
-            <MenuItem value="Span">Span</MenuItem>
+            <MenuItem value="Trace">Span</MenuItem>
           </Select>
         </FormControl>
-      </Stack>
-      <Stack direction="row" spacing={2} alignItems="center" flexWrap="wrap">
-        {isValidInputs === false && (
-          <Typography variant="caption" color="red">
-            check inputs
-          </Typography>
-        )}
-        <Button size="small" type="submit" variant="contained">
-          save
-        </Button>
-        <Button size="small" variant="outlined" onClick={handleCancel}>
-          cancel
-        </Button>
       </Stack>
     </Box>
   );
